+++ news.md
@@ -1,3 +1,10 @@
+# galamm (development versions)
+
+- Improved tests.
+- Improved error messages.
+- Argument load_var to galamm has been renamed to load_var. The old name 
+  load_var will work for a while, but a deprecation warning will be issued.
+
 # galamm 0.2.3
 
 - A bug causing heteroscedastic models with more than two groups to fail has
@@ -40,7 +47,7 @@
 
 - Fixed bug causing galamm to fail on R4.2.3.
 - Fixed memory issues in C++ code.
-- In the smooth terms sl() and t2(), the argument 'load.var' has been renamed to
+- In the smooth terms sl() and t2(), the argument 'load_var' has been renamed to
   'factor', to correspond with the remaining factor arguments.
 
 # galamm version 0.1.0